Commit the current working stage of the project.
The code needs refactoring.

Current Status:
There are two .fxml files - one for window and second for workspace.
They also have their respective controllers. The workspace
controller performs the task of adding and deleting vertex and edge.
It contains many additional method needed for the above mentioned
tasks. There is also no model layer of the project.

Refactoring requried:
Both the .fxml files will be merged into a single file. There will be
one controller. A Graph.java class will be there in the model layer
to store information related to the graph so far contructed.
The task of addition and deletion of the vertex will be delegated to
the Graph.java class which will contain methods needed for these task.

Need of refactoring:
This new structure is expected to help in the future expansion of the
project.