Merge both the controller and layout files
As the first step in adding a working model layer to the project,
both the layout files (window.fxml and workspace.fxml) have been merged
together along with their controllers (Window.java and Workspace.java).

Earlier, Workspace.java was handling all the responsibilites of
the model layer. On a small scale this was all right. Since the
project is expanding, this will not work any further.

Currently, Window.java will handle all the responsibilites of the
model layer. Soon, this will be removed once the model layer is
made active.