Add bi-directional edge feature.
On addition of reverse edge in case of directed graphs, no more new
edge will be created. The reverse arrow will be added to already
exisitng edge.

Project was also updated so on deletion of edges in case of
directed or undirected graph, correct edge will be deleted. If
edge is directed and is bidirectional, then it will be made
unidirectional first and on deleting same edge again will it be
completely deleted.

Complete testing of this feature is in process.