Do bug fix and improve code.
Bug 1: Earlier in the graphType ToggleGroup, both the options
could be deselected. This is not desired. graphType ToggleGroup
consisted of 2 ToggleButton - directed and undirected.

Fix 1: directed and undirected have been made now RadioButton
which possess the functionality that when added in a ToggleGroup,
one RadioButton must be selected. To make the RadioButton visually
appealing, the default radio-button style class have been removed.

Improvement 1: hover effect on all the buttons (toggle or radio) is
no longer managed manually. In the window.css style sheet,
toggle-button:hover is added to do this automatically.

Improvement 2: selection effect on all the buttons (toggle or radio) is
no longer managed manually. In the window.css style sheet,
toggle-button:selected is added to do this automatically for all the
ToggleButton. radio-button-custom:selected has been added as another
style class for RadioButton (directed and undirected) to override the
toggle-button:selected effect.